requirements.txt
flask==3.0.2
flask-socketio==5.3.6
fastapi==0.110.0
python-socketio==5.12.1
uvicorn==0.27.1
psutil==5.9.8
jinja2==3.1.3 